Реализован: ✅ запрос картинок из сети ✅ полноразмерный просмотр загруженной картинки ✅ функционал лайков ✅ кнопка логаута ✅ оптимизирован код
